telegram-bot/requirements.txt
aiogram==3.20.0.post0
python-telegram-bot==22.0
httpx==0.28.1
pydantic==2.11.3
pydantic-settings==2.8.1
